[ADD] Agregar Lineas a Orden de Venta
Agregado de apartado para agregar lineas a las ordenes de venta

+++ backend/src/main/java/gaseco/backend/AppWeb/AppVentas/DTO/Request/OrdenVentaRequest.java
@@ -1,14 +1,11 @@
 package gaseco.backend.AppWeb.AppVentas.DTO.Request;
 
-
-import java.time.OffsetDateTime;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.time.OffsetDateTime;
+
 
 @Data
 @Builder
@@ -16,7 +13,7 @@ import java.time.OffsetDateTime;
 @AllArgsConstructor
 public class OrdenVentaRequest {
      private String CustID;
-     private long CustNum;
+     private int CustNum;
      private String CurrencyCod;
      private String FechaR;
 }
